requirements.txt
asgiref>=3.5.2
azure-common>=1.1.28
azure-core>=1.24.1
azure-identity>=1.15.0
azure-keyvault>=4.2.0
azure-keyvault-certificates>=4.4.0
azure-keyvault-keys>=4.5.1
azure-keyvault-secrets>=4.4.0
azure-storage-blob>=12.12.0
certifi>=2022.5.18.1
cffi>=1.15.0
charset-normalizer>=2.0.12
cryptography>=37.0.2
Django>=4.0.5
django-rest-framework>=0.1.0
djangorestframework>=3.13.1
gunicorn>=20.1.0
idna>=3.3
isodate>=0.6.1
josepy>=1.13.0
mozilla-django-oidc>=2.0.0
msrest>=0.6.21
mssql-django>=1.1.3
oauthlib>=3.2.0
Pillow>=9.1.1
polars>=0.13.52
pycparser>=2.21
pyodbc>=4.0.32
python-dotenv>=0.20.0
pyOpenSSL>=22.0.0
pytz>=2022.1
requests>=2.27.1
requests-oauthlib>=1.3.1
six>=1.16.0
sqlparse>=0.4.2
typing_extensions>=4.2.0
tzdata>=2022.1
urllib3>=1.26.9
whitenoise>=6.2.0
